Move getYouTubeId to helpers
