+++ client/src/utils/helpers.js
@@ -259,3 +259,22 @@ export const createHashMap = ( array, key ) => (
     [obj[key]]: obj
   } ), {} )
 );
+
+/**
+ * Fetch id fron url. A Youtube link can either use the
+ * short form or long form so check both
+ * @param {string} url youtube share url
+ * @return youtube id
+ */
+export const getYouTubeId = ( url ) => {
+  const reShort = /https:\/\/youtu.be\/(.*)/;
+  const reLong = /https:\/\/www.youtube.com\/watch\?v=(.*)/;
+  const idShort = url.match( reShort );
+  const idLong = url.match( reLong );
+  if ( idShort ) {
+    return idShort[1];
+  } else if ( idLong ) {
+    return idLong[1];
+  }
+  return null;
+};
